Use ImageStreamTag to refer to source image
When invoked the image pruner in OpenShift automatically removes images
and other data based on a series of criteria[1]. The way the customized
Elasticsearch image was referenced did not allow the pruner to match it
up with the image tag and caused the image and its layers to be removed,
breaking Elasticsearch in the process (pods would be stuck in
ImagePullBackOff).

Using an ImageChange trigger with `automatic` set to `false` creates
a followable reference to the image tag.

[1]
https://docs.openshift.com/container-platform/3.4/admin_guide/pruning_resources.html#image-prune-conditions